+++ packages/components/src/Form/Inputs/InputSearch/InputSearch.tsx
@@ -184,7 +184,6 @@ const InputSearchLayout = forwardRef(
           <ComboboxList
             persistSelection
             windowedOptions={windowedOptions}
-            cancelClickOutside={false}
             indicator={indicator}
             width={autoResize ? 'auto' : undefined}
             aria-busy={isLoading}