+++ packages/components/src/Menu/MenuContext.ts
@@ -24,7 +24,7 @@
 
  */
 
-import { createContext, KeyboardEvent, RefObject } from 'react'
+import { createContext, KeyboardEvent } from 'react'
 
 export interface MenuContextProps {
   disabled?: boolean
@@ -34,7 +34,6 @@ export interface MenuContextProps {
   setOpen?: (isOpen: boolean) => void
   triggerElement?: HTMLElement | null
   triggerCallbackRef?: (node: HTMLElement | null) => void
-  groupedMenusRef?: RefObject<HTMLElement>
 }
 
 export interface MenuItemContextProps {